@@ -1,6 +1,7 @@
 {
   "name": "pms-console",
   "version": "1.0.0",
+  "private": true,
   "description": "Web management console for PMS",
   "main": "js/init.js",
   "scripts": {
@@ -16,20 +17,20 @@
   "author": "PonomareVlad",
   "license": "AGPL-3.0+",
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-dynamic-import-webpack": "^1.0.1",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.5",
+    "babel-plugin-dynamic-import-webpack": "^1.0.2",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-preset-env": "^1.6.1",
-    "clean-webpack-plugin": "^0.1.17",
-    "css-loader": "^0.28.7",
-    "express": "^4.16.2",
-    "file-loader": "^1.1.5",
-    "html-loader": "^0.5.1",
+    "babel-preset-env": "^1.7.0",
+    "clean-webpack-plugin": "^0.1.19",
+    "css-loader": "^0.28.11",
+    "express": "^4.16.3",
+    "file-loader": "^1.1.11",
+    "html-loader": "^0.5.5",
     "html-webpack-plugin": "^2.30.1",
-    "style-loader": "^0.19.0",
-    "webpack": "^3.8.1",
-    "webpack-dev-middleware": "^1.12.0",
-    "webpack-dev-server": "^2.9.3"
+    "style-loader": "^0.19.1",
+    "webpack": "^3.12.0",
+    "webpack-dev-middleware": "^1.12.2",
+    "webpack-dev-server": "^3.1.6"
   }
 }
